requirements.txt
# This is a comment, to show how #-prefixed lines are ignored.
# It is possible to specify requirements as plain names.
#pytest
python-gitlab
jira

# The syntax supported here is the same as that of requirement specifiers.
#docopt == 0.6.1
#requests [security] >= 2.8.1, == 2.8.* ; python_version < "2.7"
#urllib3 @ https://github.com/urllib3/urllib3/archive/refs/tags/1.26.8.zip

# It is possible to refer to other requirement files or constraints files.
#-r other-requirements.txt
#-c constraints.txt

# It is possible to refer to specific local distribution paths.
#./downloads/numpy-1.9.2-cp34-none-win32.whl

# It is possible to refer to URLs.
#http://wxpython.org/Phoenix/snapshot-builds/wxPython_Phoenix-3.0.3.dev1820+49a8884-cp34-none-win_amd64.whl